Rename BulkByScrollWireSerializingTestUtils
Rename to BulkByPaginatedSearchWireSerializingTestUtils because
reindexing now uses both scroll and point-in-time search.

Relates: elastic/elasticsearch-team#2406
