+++ server/src/test/java/org/elasticsearch/index/reindex/BulkByPaginatedSearchWireSerializingTestUtils.java
@@ -43,9 +43,9 @@ import static org.elasticsearch.xcontent.json.JsonXContent.jsonXContent;
 /**
  * Shared helpers for bulk-by-scroll wire serialization tests.
  */
-public final class BulkByScrollWireSerializingTestUtils {
+public final class BulkByPaginatedSearchWireSerializingTestUtils {
 
-    private BulkByScrollWireSerializingTestUtils() {}
+    private BulkByPaginatedSearchWireSerializingTestUtils() {}
 
     /**
      * Registry sufficient for {@link ReindexRequest}, {@link UpdateByQueryRequest}, {@link DeleteByQueryRequest},
@@ -349,7 +349,7 @@ public final class BulkByScrollWireSerializingTestUtils {
             case 13 -> mutatedRequest.setResumeInfo(
                 ESTestCase.randomValueOtherThan(
                     originalRequest.getResumeInfo().orElse(null),
-                    BulkByScrollWireSerializingTestUtils::randomResumeInfo
+                    BulkByPaginatedSearchWireSerializingTestUtils::randomResumeInfo
                 )
             );
             case 14 -> {
