+++ server/src/test/java/org/elasticsearch/index/reindex/ResumeBulkByPaginatedSearchRequestWireSerializingTests.java
@@ -19,8 +19,8 @@ import java.io.IOException;
 import java.util.Arrays;
 import java.util.Objects;
 
-import static org.elasticsearch.index.reindex.BulkByScrollWireSerializingTestUtils.reindexRequestsEqual;
-import static org.elasticsearch.index.reindex.BulkByScrollWireSerializingTestUtils.resumeInfoOptionalContentHashCode;
+import static org.elasticsearch.index.reindex.BulkByPaginatedSearchWireSerializingTestUtils.reindexRequestsEqual;
+import static org.elasticsearch.index.reindex.BulkByPaginatedSearchWireSerializingTestUtils.resumeInfoOptionalContentHashCode;
 import static org.elasticsearch.index.reindex.ReindexRequestWireSerializingTests.mutateReindexRequest;
 import static org.elasticsearch.index.reindex.ReindexRequestWireSerializingTests.newRandomReindexWireInstance;
 
@@ -29,7 +29,7 @@ public class ResumeBulkByPaginatedSearchRequestWireSerializingTests extends Abst
 
     @Override
     protected NamedWriteableRegistry getNamedWriteableRegistry() {
-        return BulkByScrollWireSerializingTestUtils.bulkScrollRequestNamedWriteableRegistry();
+        return BulkByPaginatedSearchWireSerializingTestUtils.bulkScrollRequestNamedWriteableRegistry();
     }
 
     @Override
